# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Personal Finance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a new budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetRequest"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetRequest"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all savings goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: string
        parentId:
          type: integer
        isDefault:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - color
        - icon
        - isDefault
        - createdAt
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: string
        parentId:
          type: integer
      required:
        - name
        - type
        - color
        - icon
    Budget:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        spent:
          type: number
        month:
          type: integer
        year:
          type: integer
        alertThreshold:
          type: number
        createdAt:
          type: string
        category:
          $ref: "#/components/schemas/Category"
      required:
        - id
        - categoryId
        - amount
        - spent
        - month
        - year
        - createdAt
    CreateBudgetRequest:
      type: object
      properties:
        categoryId:
          type: integer
        amount:
          type: number
        month:
          type: integer
        year:
          type: integer
        alertThreshold:
          type: number
      required:
        - categoryId
        - amount
        - month
        - year
    Goal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        targetDate:
          type: string
        color:
          type: string
        icon:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, completed, paused]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - targetAmount
        - currentAmount
        - status
        - createdAt
        - updatedAt
    CreateGoalRequest:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        targetDate:
          type: string
        color:
          type: string
        icon:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, completed, paused]
      required:
        - name
        - targetAmount
...